# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Team Titans - Free Lunch API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/organization/invite": {
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Create an organization invitation",
                "operationId": "deb4113e8c51bb3a7f4e2344436dfef0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john.doe@example.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Send email invitation from organization",
                                        "value": {
                                            "statusCode": 200,
                                            "message": "success",
                                            "data": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Send email invitation from organization",
                                        "value": {
                                            "message": "You are not authorized to perform this action!"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/withdrawal/request": {
            "get": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Get Withdrawal Requests from User",
                "operationId": "0f4064b6da1781c0d91cd7bbc590f76b",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Get Withdrawal Requests",
                                        "value": {
                                            "message": "User details fetched",
                                            "statusCode": 200,
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Withdrawal Request",
                "operationId": "c8147d0db50cb1d94b536db31fd70324",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "amount": "1000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "User Signup response",
                                        "value": {
                                            "message": "Withdrawal request created ",
                                            "statusCode": 201,
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}